+++ miniconf/src/json.rs
@@ -1,90 +1,63 @@
-//! `TreeSerialize`/`TreeDeserialize` with "JSON and `/`".
-//!
-//! Access items with `'/'` as path separator and JSON (from `serde-json-core`)
-//! as serialization/deserialization payload format.
-//!
-//! Paths used here are reciprocal to `TreeSchema::lookup::<Path<_, '/'>, _>(...)`/
-//! `TreeSchema::SCHEMA.nodes::<Path<_, '/'>>()`.
-//!
-//! ```
-//! use miniconf::{json, Tree};
-//! #[derive(Tree, Default)]
-//! struct S {
-//!     foo: u32,
-//!     bar: [u16; 2],
-//! };
-//! let mut s = S::default();
-//! json::set(&mut s, "/bar/1", b"9").unwrap();
-//! assert_eq!(s.bar[1], 9);
-//! let mut buf = [0u8; 10];
-//! let len = json::get(&mut s, "/bar/1", &mut buf[..]).unwrap();
-//! assert_eq!(&buf[..len], b"9");
-//! ```
+//! Utilities using `serde_json`
+use serde_json::value::{Serializer as ValueSerializer, Value};
 
-use serde_json_core::{de, ser};
+use crate::{Internal, IntoKeys, KeyError, Schema, SerdeError, TreeSerialize, ValueError};
 
-use crate::{IntoKeys, Path, SerdeError, TreeDeserialize, TreeSerialize};
-
-/// Update a node by path.
-///
-/// # Args
-/// * `tree` - The `TreeDeserialize` to operate on.
-/// * `path` - The path to the node. Everything before the first `'/'` is ignored.
-/// * `data` - The serialized data making up the content.
-///
-/// # Returns
-/// The number of bytes consumed from `data` or an [SerdeError].
-#[inline]
-pub fn set<'de>(
-    tree: &mut (impl TreeDeserialize<'de> + ?Sized),
-    path: &str,
-    data: &'de [u8],
-) -> Result<usize, SerdeError<de::Error>> {
-    set_by_key(tree, Path::<_, '/'>(path), data)
-}
-
-/// Retrieve a serialized value by path.
-///
-/// # Args
-/// * `tree` - The `TreeDeserialize` to operate on.
-/// * `path` - The path to the node. Everything before the first `'/'` is ignored.
-/// * `data` - The buffer to serialize the data into.
-///
-/// # Returns
-/// The number of bytes used in the `data` buffer or an [SerdeError].
-#[inline]
-pub fn get(
-    tree: &(impl TreeSerialize + ?Sized),
-    path: &str,
-    data: &mut [u8],
-) -> Result<usize, SerdeError<ser::Error>> {
-    get_by_key(tree, Path::<_, '/'>(path), data)
-}
-
-/// Update a node by key.
-///
-/// # Returns
-/// The number of bytes consumed from `data` or an [SerdeError].
-pub fn set_by_key<'de>(
-    tree: &mut (impl TreeDeserialize<'de> + ?Sized),
-    keys: impl IntoKeys,
-    data: &'de [u8],
-) -> Result<usize, SerdeError<de::Error>> {
-    let mut de = de::Deserializer::new(data, None);
-    tree.deserialize_by_key(keys.into_keys(), &mut de)?;
-    de.end().map_err(SerdeError::Finalization)
-}
-
-/// Retrieve a serialized value by key.
-///
-/// # Returns
-/// The number of bytes used in the `data` buffer or an [SerdeError].
-pub fn get_by_key(
-    tree: &(impl TreeSerialize + ?Sized),
-    keys: impl IntoKeys,
-    data: &mut [u8],
-) -> Result<usize, SerdeError<ser::Error>> {
-    let mut ser = ser::Serializer::new(data);
-    tree.serialize_by_key(keys.into_keys(), &mut ser)?;
-    Ok(ser.end())
+/// Serialize a TreeSerialize into a JSON Value
+pub fn to_json_value<T: TreeSerialize>(
+    value: &T,
+) -> Result<Value, SerdeError<<ValueSerializer as serde::Serializer>::Error>> {
+    fn visit<T: TreeSerialize>(
+        idx: &mut [usize],
+        depth: usize,
+        schema: &Schema,
+        value: &T,
+    ) -> Result<Value, SerdeError<<ValueSerializer as serde::Serializer>::Error>> {
+        match value.serialize_by_key((&idx[..depth]).into_keys(), ValueSerializer) {
+            Ok(v) => Ok(v),
+            Err(SerdeError::Value(ValueError::Absent)) => {
+                Ok(Value::String("__tree-absent__".to_string()))
+            }
+            Err(SerdeError::Value(ValueError::Access(_msg))) => {
+                Ok(Value::String("__tree-access__".to_string()))
+            }
+            Err(SerdeError::Value(ValueError::Key(KeyError::TooShort))) => {
+                Ok(match schema.internal.as_ref().unwrap() {
+                    Internal::Homogeneous(h) => Value::Array(
+                        (0..h.len.get())
+                            .map(|i| {
+                                idx[depth] = i;
+                                visit(idx, depth + 1, h.schema, value)
+                            })
+                            .collect::<Result<_, _>>()?,
+                    ),
+                    Internal::Named(n) => Value::Object(
+                        n.iter()
+                            .enumerate()
+                            .map(|(i, n)| {
+                                idx[depth] = i;
+                                Ok((n.name.to_string(), visit(idx, depth + 1, n.schema, value)?))
+                            })
+                            .collect::<Result<_, SerdeError<_>>>()?,
+                    ),
+                    Internal::Numbered(n) => Value::Array(
+                        n.iter()
+                            .enumerate()
+                            .map(|(i, n)| {
+                                idx[depth] = i;
+                                visit(idx, depth + 1, n.schema, value)
+                            })
+                            .collect::<Result<_, _>>()?,
+                    ),
+                })
+            }
+            Err(err) => Err(err),
+        }
+    }
+    visit(
+        &mut vec![0; T::SCHEMA.shape().max_depth],
+        0,
+        T::SCHEMA,
+        value,
+    )
 }