+++ miniconf/tests/structs.rs
@@ -1,6 +1,6 @@
 use miniconf::{
     Deserialize, IntoKeys, Leaf, Serialize, Shape, Tree, TreeAny, TreeDeserialize, TreeSchema,
-    TreeSerialize, ValueError, json,
+    TreeSerialize, ValueError, json_core,
 };
 
 mod common;
@@ -24,7 +24,7 @@ fn structs() {
     let mut settings = Settings::default();
 
     // Inner settings structure is atomic, so cannot be set.
-    assert!(json::set(&mut settings, "/c/a", b"4").is_err());
+    assert!(json_core::set(&mut settings, "/c/a", b"4").is_err());
 
     // Inner settings can be updated atomically.
     set_get(&mut settings, "/c", b"{\"a\":5}");
@@ -33,7 +33,7 @@ fn structs() {
     set_get(&mut settings, "/d/a", b"3");
 
     // It is not allowed to set a non-terminal node.
-    assert!(json::set(&mut settings, "/d", b"{\"a\": 5").is_err());
+    assert!(json_core::set(&mut settings, "/d", b"{\"a\": 5").is_err());
 
     assert_eq!(*settings.c, Inner { a: 5 });
     assert_eq!(settings.d, Inner { a: 3 });
@@ -70,8 +70,8 @@ fn tuple_struct() {
     assert_eq!(s.0, 2);
     set_get(&mut s, "/1", br#"3.0"#);
     assert_eq!(s.1, 3.0);
-    json::set(&mut s, "/2", b"3.0").unwrap_err();
-    json::set(&mut s, "/foo", b"3.0").unwrap_err();
+    json_core::set(&mut s, "/2", b"3.0").unwrap_err();
+    json_core::set(&mut s, "/foo", b"3.0").unwrap_err();
 
     assert_eq!(paths::<Settings, 1>(), ["/0", "/1"]);
 }