+++ miniconf_mqtt/src/lib.rs
@@ -11,7 +11,7 @@ use heapless::{String, Vec};
 use log::{error, info, warn};
 use miniconf::{
     DescendError, IntoKeys, KeyError, Keys, NodeIter, Path, Schema, SerdeError, Track,
-    TreeDeserializeOwned, TreeSchema, TreeSerialize, ValueError, json,
+    TreeDeserializeOwned, TreeSchema, TreeSerialize, ValueError, json_core,
 };
 pub use minimq;
 use minimq::{
@@ -459,7 +459,7 @@ where
 
             let props = [ResponseCode::Ok.into()];
             let mut response = Publication::new(&topic, |buf: &mut [u8]| {
-                track_depth(&path, |k| json::get_by_key(settings, k, buf))
+                track_depth(&path, |k| json_core::get_by_key(settings, k, buf))
             })
             .properties(&props)
             .qos(QoS::AtLeastOnce);
@@ -538,7 +538,7 @@ where
                 // Try a Get assuming a leaf node
                 if let Err(err) = client.publish(
                     Publication::respond(Some(topic), properties, |buf: &mut [u8]| {
-                        track_depth(path, |k| json::get_by_key(settings, k, buf))
+                        track_depth(path, |k| json_core::get_by_key(settings, k, buf))
                     })
                     .unwrap()
                     .properties(&[ResponseCode::Ok.into()])
@@ -584,7 +584,7 @@ where
                 State::Unchanged
             } else {
                 // Set
-                match track_depth(path, |k| json::set_by_key(settings, k, payload)) {
+                match track_depth(path, |k| json_core::set_by_key(settings, k, payload)) {
                     Err(err) => {
                         Self::respond(err, ResponseCode::Error, properties, client);
                         State::Unchanged